server-side fetching for all pages in settings/admin
