+++ apps/web/app/future/settings/(admin)/admin/flags/page.tsx
@@ -1,7 +1,9 @@
 import { _generateMetadata } from "app/_utils";
+import { notFound } from "next/navigation";
 
 import { FlagListingView } from "@calcom/features/flags/pages/flag-listing-view";
 import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
+import { FeatureFlagRepository } from "@calcom/lib/server/repository/featureFlag";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -10,11 +12,16 @@ export const generateMetadata = async () =>
   );
 
 const Page = async () => {
-  return (
-    <SettingsHeader title="Feature Flags" description="Here you can toggle your Cal.com instance features.">
-      <FlagListingView />
-    </SettingsHeader>
-  );
+  try {
+    const featureFlags = await FeatureFlagRepository.getFeatureFlags();
+    return (
+      <SettingsHeader title="Feature Flags" description="Here you can toggle your Cal.com instance features.">
+        <FlagListingView featureFlags={featureFlags} />
+      </SettingsHeader>
+    );
+  } catch (error) {
+    notFound();
+  }
 };
 
 export default Page;
