+++ apps/web/app/future/settings/(admin)/admin/page.tsx
@@ -1,8 +1,4 @@
-import LegacyPage from "@pages/settings/admin/index";
 import { _generateMetadata } from "app/_utils";
-import { WithLayout } from "app/layoutHOC";
-
-import { getLayout } from "@components/auth/layouts/AdminLayoutAppDir";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -10,4 +6,7 @@ export const generateMetadata = async () =>
     () => "admin_description"
   );
 
-export default WithLayout({ getServerLayout: getLayout, Page: LegacyPage })<"P">;
+const Page = () => {
+  return <h1>Admin index</h1>;
+};
+export default Page;
