+++ apps/web/pages/settings/admin/flags.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import NoSSR from "@calcom/core/components/NoSSR";
 import { FlagListingView } from "@calcom/features/flags/pages/flag-listing-view";
 import { Meta } from "@calcom/ui";
 
@@ -7,10 +8,12 @@ import PageWrapper from "@components/PageWrapper";
 import { getLayout } from "@components/auth/layouts/AdminLayout";
 
 const FlagsPage = () => (
-  <div>
-    <Meta title="Feature Flags" description="Here you can toggle your Cal.com instance features." />
-    <FlagListingView />
-  </div>
+  <NoSSR>
+    <div>
+      <Meta title="Feature Flags" description="Here you can toggle your Cal.com instance features." />
+      <FlagListingView />
+    </div>
+  </NoSSR>
 );
 
 FlagsPage.getLayout = getLayout;
