+++ packages/features/flags/server/router.ts
@@ -1,15 +1,12 @@
+import { FeatureFlagRepository } from "@calcom/lib/server/repository/featureFlag";
 import publicProcedure from "@calcom/trpc/server/procedures/publicProcedure";
 import { router } from "@calcom/trpc/server/trpc";
 
 import { map } from "./procedures/map";
 
 export const featureFlagRouter = router({
-  list: publicProcedure.query(async ({ ctx }) => {
-    const { prisma } = ctx;
-    return prisma.feature.findMany({
-      orderBy: { slug: "asc" },
-      cacheStrategy: { swr: 300, ttl: 300 },
-    });
+  list: publicProcedure.query(async () => {
+    return await FeatureFlagRepository.getFeatureFlags();
   }),
   map,
 });
